+++ packages/shopify-codemod/test/fixtures/implicit-coercion-to-explicit/basic.output.js
@@ -1,0 +1,4 @@
+Boolean(foo);
+Boolean(!foo);
+Number(foo);
+++foo;