+++ packages/shopify-codemod/test/transforms/implicit-coercion-to-explicit.test.js
@@ -1,0 +1,8 @@
+import 'test-helper';
+import transform from 'implicit-coercion-to-explicit';
+
+describe('implicitCoercionToExplicit', () => {
+  it('transforms the most offensive implicit coercion to be explicit', () => {
+    expect(transform).to.transform('implicit-coercion-to-explicit/basic');
+  });
+});